+++ packages/core/chat/store.test.ts
@@ -97,22 +97,24 @@ describe("chat store — draft attachments", () => {
 });
 
 describe("chat store — floating window preference", () => {
-  it("defaults ON when no preference is stored", () => {
+  it("defaults OFF when no preference is stored (opt-in)", () => {
     const store = createChatStore({ storage: memStorage() });
-    expect(store.getState().floatingChatEnabled).toBe(true);
-  });
-
-  it("honours an explicit stored 'false' preference", () => {
-    const storage = memStorage();
-    storage.setItem("multica:chat:floatingChatEnabled", "false");
-    const store = createChatStore({ storage });
     expect(store.getState().floatingChatEnabled).toBe(false);
   });
 
-  it("persists the choice and collapses an open overlay when disabled", () => {
+  it("honours an explicit stored 'true' preference", () => {
     const storage = memStorage();
+    storage.setItem("multica:chat:floatingChatEnabled", "true");
+    const store = createChatStore({ storage });
+    expect(store.getState().floatingChatEnabled).toBe(true);
+  });
+
+  it("persists an enable, then collapses an open overlay when disabled again", () => {
+    const storage = memStorage();
+    storage.setItem("multica:chat:floatingChatEnabled", "true");
     storage.setItem("multica:chat:isOpen", "true");
     const store = createChatStore({ storage });
+    expect(store.getState().floatingChatEnabled).toBe(true);
     expect(store.getState().isOpen).toBe(true);
 
     store.getState().setFloatingChatEnabled(false);
@@ -126,5 +128,6 @@ describe("chat store — floating window preference", () => {
 
     store.getState().setFloatingChatEnabled(true);
     expect(store.getState().floatingChatEnabled).toBe(true);
+    expect(storage.getItem("multica:chat:floatingChatEnabled")).toBe("true");
   });
 });
