+++ packages/core/chat/store.ts
@@ -40,7 +40,8 @@ const OPEN_KEY = "multica:chat:isOpen";
  * Whether the floating chat window (FAB + overlay) is available at all,
  * persisted globally like OPEN_KEY. This is the Settings → Chat preference:
  * when off, the FAB/overlay never mount and Chat lives only in its tab.
- * Missing key = default ON, so existing users keep the floating entry point.
+ * Missing key = default OFF — the floating window is opt-in, enabled from
+ * the Settings → Chat tab.
  */
 const FLOATING_KEY = "multica:chat:floatingChatEnabled";
 
@@ -178,8 +179,9 @@ export function createChatStore(options: ChatStoreOptions) {
   const storedOpen = storage.getItem(OPEN_KEY);
   const initialIsOpen = storedOpen === "true";
 
-  // Default ON: only an explicit "false" disables the floating window.
-  const initialFloatingEnabled = storage.getItem(FLOATING_KEY) !== "false";
+  // Default OFF: the floating window is opt-in — only an explicit "true"
+  // (set from the Settings → Chat tab) enables it.
+  const initialFloatingEnabled = storage.getItem(FLOATING_KEY) === "true";
 
   const store = create<ChatState>((set, get) => ({
     isOpen: initialIsOpen,
