+++ packages/views/locales/ko/settings.json
@@ -1,10 +1,5 @@
 {
   "preferences": {
-    "chat": {
-      "title": "채팅",
-      "floating_label": "플로팅 채팅 창",
-      "floating_hint": "모든 페이지 모서리에 플로팅 채팅 버튼을 표시합니다. 끄면 채팅은 전용 탭에서만 사용할 수 있습니다."
-    },
     "theme": {
       "title": "테마",
       "light": "라이트",
@@ -26,6 +21,11 @@
       "sync_failed": "시간대 설정을 저장하지 못했습니다."
     }
   },
+  "chat": {
+    "floating_title": "플로팅 창",
+    "floating_label": "플로팅 채팅 창",
+    "floating_hint": "모든 페이지 모서리에 플로팅 채팅 버튼을 표시합니다. 기본값은 꺼짐 — 켜면 어디서나 채팅을 불러올 수 있고, 꺼면 채팅은 이 탭에서만 사용할 수 있습니다."
+  },
   "page": {
     "title": "설정",
     "my_account": "내 계정",
@@ -33,6 +33,7 @@
     "tabs": {
       "profile": "프로필",
       "preferences": "환경설정",
+      "chat": "채팅",
       "notifications": "알림",
       "tokens": "API 토큰",
       "general": "일반",
